pyproject.toml
[tool.poetry]
name = "liboard"
version = "2.0.0"
description = "Python module for LiBoard"
authors = ["Philipp Leclercq <liboard@mailbox.org>"]
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.9"
python-chess = "^1.999"
pyserial = "^3.5"
bitstring = "^3.1.9"
PyAutoGUI = "^0.9.53"
httpx = "^0.21.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
flake8-docstrings = "^1.6.0"

[tool.autopep8]
max-line-length = 99

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
